Network: fix int overflow in DAG blocks structure

+++ network/proto/blocks.go
@@ -89,7 +89,7 @@ type head struct {
 	// distance contains number of blocks to the next TX that refers to this TX as 'prev'. If the next TX is in the next
 	// block, distance will be 1. If `distance is 0, it means the next TX is in the same block. It is initialized
 	// to math.MaxInt64, meaning no next TX.
-	distance    int
+	distance    int64
 	signingTime time.Time
 	// blockDate contains the signing time at start of the block
 	blockDate time.Time
@@ -165,7 +165,7 @@ func (blx *trackingDAGBlocks) internalAddTransaction(tx dag.Transaction) {
 	for i := 0; i <= blockIdx; i++ {
 		for _, currPrev := range tx.Previous() {
 			if head, ok := blx.blocks[i].heads[currPrev]; ok {
-				newDistance := int(txBlockDate.Sub(head.blockDate).Hours() / 24)
+				newDistance := int64(txBlockDate.Sub(head.blockDate).Hours() / 24)
 				if newDistance < head.distance {
 					head.distance = newDistance
 				}
